Extensions files naming problem solved Folders organisation corrected
