+++ TheMET/TheMET/Extentions/UILabel+apply.swift
@@ -1,0 +1,29 @@
+//
+//  UILabel+apply.swift
+//  TheMET
+//
+//  Created by Анна Ситникова on 11/05/2023.
+//
+
+import Foundation
+import UIKit
+
+extension UILabel {
+    func apply(font: String, color: UIColor?, fontSize: Double, title: String) {
+        let paragraphStyle = NSMutableParagraphStyle()
+        paragraphStyle.lineHeightMultiple = 1.3
+        paragraphStyle.alignment = .center
+        var labelAttributes: [NSAttributedString.Key: Any] = [ : ]
+        let labelfontName = font
+        if let labelFont: UIFont = UIFont(name: labelfontName, size: 18),
+           let color: UIColor = color {
+            labelAttributes[.font] = labelFont
+            labelAttributes[.foregroundColor] = color
+            if labelfontName == "Quattrocento" {
+                labelAttributes[.paragraphStyle] = paragraphStyle
+            }
+        }
+        let attribuesString = NSMutableAttributedString(string: title, attributes: labelAttributes)
+        self.attributedText = attribuesString
+    }
+}
